Basic Structuring and fixes
- Add a readme.
- Add CI tests.
- Removed main.rs.
- Use lib.rs to maintain library structure.
- Fix clippy warnings and cargo fmt fixes.
