+++ .github/workflows/test.yaml
@@ -1,0 +1,42 @@
+on:
+  push:
+    branches:
+    - master
+  pull_request:
+
+name: test
+
+jobs:
+  test_with_codecov:
+    name: Run tests with coverage reporting
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v3
+
+      - name: Set default toolchain
+        run: rustup default nightly
+      - name: Set profile
+        run: rustup set profile minimal
+
+      # Pin grcov to v0.8.2 because of build failure at 0.8.3
+      - name: Install grcov
+        run: cargo install grcov --force --version 0.8.2
+
+      # Tests are run with code coverage support
+      - name: Run cargo test
+        env:
+          CARGO_INCREMENTAL: '0'
+          RUSTFLAGS: '-Zprofile -Ccodegen-units=1 -Cinline-threshold=0 -Clink-dead-code -Coverflow-checks=off'
+          RUSTDOCFLAGS: '-Zprofile -Ccodegen-units=1 -Cinline-threshold=0 -Clink-dead-code -Coverflow-checks=off'
+        run: cargo test -- --nocapture
+      - id: coverage
+        name: Generate coverage
+        uses: actions-rs/grcov@v0.1.5
+      
+      # Upload coverage report
+      - name: Upload coverage to Codecov
+        uses: codecov/codecov-action@v1
+        with:
+          file: ${{ steps.coverage.outputs.report }}
+          directory: ./coverage/reports/
