+++ README.md
@@ -1,0 +1,3 @@
+# Programming Bitcoin In Rust
+
+This repo contains the rust version of the [Programming Bitcoin](https://github.com/jimmysong/programmingbitcoin) exercises.
